add overflow container
overflow container allow child widget to take as much space as they want
without shrinking its content to fit a limited space.
TODO: add scrolling behaviors to access hidden parts of the child widget.

this placement requires a modification in wibox.hierarchy so that mouse
events wont be processed for the hidden part of the child widget.
check awesomeww pr-3309 for details.